Implement SSR for SlideOut
