+++ js/ui/slide_out.js
@@ -4,7 +4,6 @@ var $ = require("../core/renderer"),
     getPublicElement = require("../core/utils/dom").getPublicElement,
     noop = require("../core/utils/common").noop,
     isDefined = require("../core/utils/type").isDefined,
-    windowUtils = require("../core/utils/window"),
     registerComponent = require("../core/component_registrator"),
     extend = require("../core/utils/extend").extend,
     map = require("../core/utils/iterator").map,
@@ -309,7 +308,7 @@ var SlideOut = CollectionWidget.inherit({
     },
 
     _renderContentTemplate: function() {
-        if(isDefined(this._singleContent) || !windowUtils.hasWindow()) {
+        if(isDefined(this._singleContent)) {
             return;
         }
