+++ testing/tests/DevExpress.serverSide/slideOut.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets/slideOut.markup.tests.js");
